+++ src/components/common/error-boundary/error-indicator.tsx
@@ -3,8 +3,8 @@ import { useHistory } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import Button from '../button/button';
-import { BASE_URL } from '../../constants/url.constants';
-import errorImage from '../../assets/error.png';
+import { BASE_URL } from '../../../constants/url.constants';
+import errorImage from '../../../assets/error.png';
 
 const Container = styled.div`
   position: absolute;
