+++ src/redux/selectors/chat.selectors.ts
@@ -1,5 +1,5 @@
 import { CHAT_NAMES, ChatStateInterface } from '../../constants/chat.constants';
-import { MessageInterface } from '../../components/messages/message/message';
+import { MessageInterface } from '../../components/common/messages/message/message';
 
 export const selectChat = ({ chat }): ChatStateInterface => chat;
 export const selectMessages = ({ chat }): MessageInterface[] => chat[CHAT_NAMES.messages];
